+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @versaprotocol/react
 
+## 1.5.0
+
+### Minor Changes
+
+- Update in accordance with schema version 1.5.1, accepting omitted keys as equivalent to null keys
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/schema@1.5.0
+  - @versaprotocol/belt@1.5.0
+
 ## 1.4.1
 
 ### Patch Changes