+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @versaprotocol/schema
 
+## 1.5.0
+
+### Minor Changes
+
+- Update in accordance with schema version 1.5.1, accepting omitted keys as equivalent to null keys
+
 ## 1.4.0
 
 ### Minor Changes